+++ Sources/WordPressSharedObjC/Logging/WPSharedLogging.m
@@ -1,0 +1,15 @@
+#import "WPSharedLogging.h"
+
+#if SWIFT_PACKAGE
+#import "../Private/WPSharedLoggingPrivate.h"
+#else
+#import "WPSharedLoggingPrivate.h"
+#endif
+
+DDLogLevel WPSharedGetLoggingLevel() {
+    return ddLogLevel;
+}
+
+void WPSharedSetLoggingLevel(DDLogLevel level) {
+    ddLogLevel = level;
+}