+++ Sources/WordPressSharedObjC/include/NSString+XMLExtensions.h
@@ -1,0 +1,1 @@
+../Utility/NSString+XMLExtensions.h